Refactor class-beans-post-meta unit tests for WPCS
