+++ tests/phpunit/integration/api/post-meta/class-beans-post-meta.php
